Data.Mutable.Class

Description

Abstract over different types for mutable references of values.

Documentation

class Monad m => Mutable m a where

Minimal complete definition

Nothing

Associated Types

type Ref m a = (v :: Type) | v -> a

type Ref m a = MutVar (PrimState m) a

Methods

thawRef :: a -> m (Ref m a)

default thawRef :: DefaultMutable m a (Ref m a) => a -> m (Ref m a)

freezeRef :: Ref m a -> m a

default freezeRef :: DefaultMutable m a (Ref m a) => Ref m a -> m a

copyRef :: Ref m a -> a -> m ()

default copyRef :: DefaultMutable m a (Ref m a) => Ref m a -> a -> m ()

modifyRef :: Ref m a -> (a -> a) -> m ()

Apply a pure function on an immutable value onto a value stored in a mutable reference.

modifyRef' :: Ref m a -> (a -> a) -> m ()

modifyRef, but forces the result before storing it back in the reference.

updateRef :: Ref m a -> (a -> (a, b)) -> m b

Apply a pure function on an immutable value onto a value stored in a mutable reference, returning a result value from that function.

updateRef' :: Ref m a -> (a -> (a, b)) -> m b

updateRef, but forces the updated value before storing it back in the reference.

newtype MutRef a

Newtype wrapper that can provide any type with a Mutable instance. Can be useful for avoiding orphan instances.

Constructors

MutRef

Fields

runMutRef :: a

newtype RefFor m a

Constructors

RefFor

Fields

getRefFor :: Ref m a

class DefaultMutable m a r where

Methods

defaultThawRef :: a -> m r

defaultFreezeRef :: r -> m a

defaultCopyRef :: r -> a -> m ()

Instances

class Monad m => GMutable m f

Minimal complete definition

gThawRef_, gFreezeRef_, gCopyRef_

type family GRef_ m f = (u :: Type -> Type) | u -> f

newtype GRef m a

Constructors

GRef

Fields

unGRef :: GRef_ m (Rep a) ()

gThawRef :: (Generic a, GMutable m (Rep a)) => a -> m (GRef m a)

gFreezeRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> m a

gCopyRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> a -> m ()

thawHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z Identity -> m (z (RefFor m))

freezeHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z (RefFor m) -> m (z Identity)

copyHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z (RefFor m) -> z Identity -> m ()

newtype RecRef m f a

Constructors

RecRef

Fields

recRef :: Ref m (f a)

ReMutable

reMutable :: forall m n a r. (Mutable m a, Monad n) => (forall x. m x -> n x) -> (Mutable n a => r) -> r

If you can provice a natural transformation from m to n, you should be able to use a value as if it had Mutable n a if you have Mutable m a.

reMutableConstraint :: forall m n a. (Mutable m a, Monad n) => (forall x. m x -> n x) -> Mutable m a :- Mutable n a

If you can provice a natural transformation from m to n, then Mutable m a should also imply Mutable n a.

newtype ReMutable (s :: Type) m a

Constructors

ReMutable a

newtype ReMutableTrans m n

Constructors

RMT

Fields

runRMT :: forall x. m x -> n x