Data.Mutable.Instances

Description

Provides Ref instances for various data types.

Instances

data ListRefCell m a

Single linked list cell

Constructors

MutNil

MutCons (Ref m a) (Ref m [a])

unconsListRef :: PrimMonad m => Ref m [a] -> m (ListRefCell m a)

Uncons mutable linked list into a ListRefCell.

consListRef :: PrimMonad m => ListRefCell m a -> m (Ref m [a])

Cons the contents of a ListRefCell into a mutable linked list.

newtype RecRef m f a

Ref for components in a vinyl Rec.

Constructors

RecRef

Fields

getRecRef :: Ref m (f a)

Generic

newtype GRef m a

Automatically generate a piecewise mutable reference for any Generic instance.

-- | any Generic instance
data Foo = Foo { fInt :: Int, fDouble :: Double }
  deriving (Generic, Show)

instance Mutable m Foo where
    type Ref m Foo = GRef m Foo

ghci> r <- thawRef (Foo 3 4.5)
ghci> freezeRef r
Foo 3 4.5
ghci> freezePart (fieldMut #fInt) r
3
ghci> copyPart (fieldMut #fDouble) 1.23
ghci> freezeRef r
Foo 3 1.23

Note that this is basically just a bunch of tupled refs for a product type. For a sum type (with multiple constructors), an extra layer of indirection is added to account for the dynamically changable shape.

See fieldMut/posMut for ways to inspect and mutate the internals of this type (as demonstrated above).

Constructors

GRef

Fields

unGRef :: GRef_ m (Rep a) ()

gThawRef :: (Generic a, GMutable m (Rep a)) => a -> m (GRef m a)

Default thawRef for GRef.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with GRef as the Ref. However, it can be useful if you are using a GRef m a just as a normal data type, independent of the Ref class. See documentation for GRef for more information.

gFreezeRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> m a

Default freezeRef for GRef.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with GRef as the Ref. However, it can be useful if you are using a GRef m a just as a normal data type, independent of the Ref class. See documentation for GRef for more information.

gCopyRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> a -> m ()

Default copyRef for GRef.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with GRef as the Ref. However, it can be useful if you are using a GRef m a just as a normal data type, independent of the Ref class. See documentation for GRef for more information.

class Monad m => GMutable m f

Class for automatic generation of Ref for Generic instances. See GRef for more information.

Minimal complete definition

gThawRef_, gFreezeRef_, gCopyRef_

Associated Types

type GRef_ m f = (u :: k -> Type) | u -> f

Higher-Kinded Data Pattern

thawHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z Identity -> m (z (RefFor m))

Default copyRef for the higher-kinded data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with z (RefFor m) as the Ref. However, it can be useful if you are using a z (RefFor m) just as a normal data type, independent of the Ref class. See documentation for Mutable for more information.

freezeHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z (RefFor m) -> m (z Identity)

Default freezeRef for the higher-kinded data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with z (RefFor m) as the Ref. However, it can be useful if you are using a z (RefFor m) just as a normal data type, independent of the Ref class. See documentation for Mutable for more information.

copyHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z (RefFor m) -> z Identity -> m ()

Default copyRef for the higher-kinded data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with z (RefFor m) as the Ref. However, it can be useful if you are using a z (RefFor m) just as a normal data type, independent of the Ref class. See documentation for Mutable for more information.

Coercible

newtype CoerceRef m s a

A Ref that works by using the Mutable instance of an equivalent type. This is useful for newtype wrappers, so you can use the underlying data type's Mutable instance.

newtype MyVec = MyVec (Vector Double)

instance Mutable m MyVec where
    type Ref m MyVec = CoerceRef m s (Vector Double)

The Ref m MyVec uses the a MVector Double under the hood.

It's essentially a special case of GRef for newtypes.

Constructors

CoerceRef

Fields

getCoerceRef :: Ref m a

thawCoerce :: (Coercible s a, Mutable m a) => s -> m (CoerceRef m s a)

freezeCoerce :: (Coercible s a, Mutable m a) => CoerceRef m s a -> m s

copyCoerce :: (Coercible s a, Mutable m a) => CoerceRef m s a -> s -> m ()

Traversable

newtype TraverseRef m f a

A Ref that works for any instance of Traversable.

Copying and modifying semantics can be a bit funky.

If copying a shorter item into a longer item ref, the "leftovers" items in the longer item are unchanged.

If copying a longer item into a shorter item ref, the leftover items are unchanged.

ghci> r <- thawTraverse [1..10]
ghci> copyTraverse r [0,0,0,0]
ghci> freezeTraverse r
[0,0,0,0,5,6,7,8,9,10]
ghci> copyTraverse r [20..50]
ghci> freezeTraverse r
[20,21,22,23,24,25,26,27,28,29]

Constructors

TraverseRef

Fields

getTraverseRef :: f (Ref m a)

thawTraverse :: (Traversable f, Mutable m a) => f a -> m (TraverseRef m f a)

freezeTraverse :: (Traversable f, Mutable m a) => TraverseRef m f a -> m (f a)

copyTraverse :: (Traversable f, Mutable m a) => TraverseRef m f a -> f a -> m ()

Instances for Generics combinators themselves

newtype GMutableRef m f a

A Ref for instances of GMutable, which are the GHC.Generics combinators.

Constructors

GMutableRef

Fields

getGMutableRef :: GRef_ m f a

thawGMutableRef :: GMutable m f => f a -> m (GMutableRef m f a)

freezeGMutableRef :: GMutable m f => GMutableRef m f a -> m (f a)

copyGMutableRef :: GMutable m f => GMutableRef m f a -> f a -> m ()

Orphan instances

Monad m => Mutable m ()

Instance details

Associated Types

type Ref m () = (v :: Type)

Methods

thawRef :: () -> m (Ref m ())

freezeRef :: Ref m () -> m ()

copyRef :: Ref m () -> () -> m ()

Monad m => Mutable m Void

Instance details

Associated Types

type Ref m Void = (v :: Type)

Methods

thawRef :: Void -> m (Ref m Void)

freezeRef :: Ref m Void -> m Void

copyRef :: Ref m Void -> Void -> m ()

PrimMonad m => Mutable m CDouble

Instance details

Associated Types

type Ref m CDouble = (v :: Type)

Methods

thawRef :: CDouble -> m (Ref m CDouble)

freezeRef :: Ref m CDouble -> m CDouble

copyRef :: Ref m CDouble -> CDouble -> m ()

PrimMonad m => Mutable m CFloat

Instance details

Associated Types

type Ref m CFloat = (v :: Type)

Methods

thawRef :: CFloat -> m (Ref m CFloat)

freezeRef :: Ref m CFloat -> m CFloat

copyRef :: Ref m CFloat -> CFloat -> m ()

PrimMonad m => Mutable m CSUSeconds

Instance details

Associated Types

type Ref m CSUSeconds = (v :: Type)

Methods

thawRef :: CSUSeconds -> m (Ref m CSUSeconds)

freezeRef :: Ref m CSUSeconds -> m CSUSeconds

copyRef :: Ref m CSUSeconds -> CSUSeconds -> m ()

PrimMonad m => Mutable m CUSeconds

Instance details

Associated Types

type Ref m CUSeconds = (v :: Type)

Methods

thawRef :: CUSeconds -> m (Ref m CUSeconds)

freezeRef :: Ref m CUSeconds -> m CUSeconds

copyRef :: Ref m CUSeconds -> CUSeconds -> m ()

PrimMonad m => Mutable m CTime

Instance details

Associated Types

type Ref m CTime = (v :: Type)

Methods

thawRef :: CTime -> m (Ref m CTime)

freezeRef :: Ref m CTime -> m CTime

copyRef :: Ref m CTime -> CTime -> m ()

PrimMonad m => Mutable m CClock

Instance details

Associated Types

type Ref m CClock = (v :: Type)

Methods

thawRef :: CClock -> m (Ref m CClock)

freezeRef :: Ref m CClock -> m CClock

copyRef :: Ref m CClock -> CClock -> m ()

PrimMonad m => Mutable m CUIntMax

Instance details

Associated Types

type Ref m CUIntMax = (v :: Type)

Methods

thawRef :: CUIntMax -> m (Ref m CUIntMax)

freezeRef :: Ref m CUIntMax -> m CUIntMax

copyRef :: Ref m CUIntMax -> CUIntMax -> m ()

PrimMonad m => Mutable m CIntMax

Instance details

Associated Types

type Ref m CIntMax = (v :: Type)

Methods

thawRef :: CIntMax -> m (Ref m CIntMax)

freezeRef :: Ref m CIntMax -> m CIntMax

copyRef :: Ref m CIntMax -> CIntMax -> m ()

PrimMonad m => Mutable m CUIntPtr

Instance details

Associated Types

type Ref m CUIntPtr = (v :: Type)

Methods

thawRef :: CUIntPtr -> m (Ref m CUIntPtr)

freezeRef :: Ref m CUIntPtr -> m CUIntPtr

copyRef :: Ref m CUIntPtr -> CUIntPtr -> m ()

PrimMonad m => Mutable m CIntPtr

Instance details

Associated Types

type Ref m CIntPtr = (v :: Type)

Methods

thawRef :: CIntPtr -> m (Ref m CIntPtr)

freezeRef :: Ref m CIntPtr -> m CIntPtr

copyRef :: Ref m CIntPtr -> CIntPtr -> m ()

PrimMonad m => Mutable m CBool

Instance details

Associated Types

type Ref m CBool = (v :: Type)

Methods

thawRef :: CBool -> m (Ref m CBool)

freezeRef :: Ref m CBool -> m CBool

copyRef :: Ref m CBool -> CBool -> m ()

PrimMonad m => Mutable m CULLong

Instance details

Associated Types

type Ref m CULLong = (v :: Type)

Methods

thawRef :: CULLong -> m (Ref m CULLong)

freezeRef :: Ref m CULLong -> m CULLong

copyRef :: Ref m CULLong -> CULLong -> m ()

PrimMonad m => Mutable m CLLong

Instance details

Associated Types

type Ref m CLLong = (v :: Type)

Methods

thawRef :: CLLong -> m (Ref m CLLong)

freezeRef :: Ref m CLLong -> m CLLong

copyRef :: Ref m CLLong -> CLLong -> m ()

PrimMonad m => Mutable m CSigAtomic

Instance details

Associated Types

type Ref m CSigAtomic = (v :: Type)

Methods

thawRef :: CSigAtomic -> m (Ref m CSigAtomic)

freezeRef :: Ref m CSigAtomic -> m CSigAtomic

copyRef :: Ref m CSigAtomic -> CSigAtomic -> m ()

PrimMonad m => Mutable m CWchar

Instance details

Associated Types

type Ref m CWchar = (v :: Type)

Methods

thawRef :: CWchar -> m (Ref m CWchar)

freezeRef :: Ref m CWchar -> m CWchar

copyRef :: Ref m CWchar -> CWchar -> m ()

PrimMonad m => Mutable m CSize

Instance details

Associated Types

type Ref m CSize = (v :: Type)

Methods

thawRef :: CSize -> m (Ref m CSize)

freezeRef :: Ref m CSize -> m CSize

copyRef :: Ref m CSize -> CSize -> m ()

PrimMonad m => Mutable m CPtrdiff

Instance details

Associated Types

type Ref m CPtrdiff = (v :: Type)

Methods

thawRef :: CPtrdiff -> m (Ref m CPtrdiff)

freezeRef :: Ref m CPtrdiff -> m CPtrdiff

copyRef :: Ref m CPtrdiff -> CPtrdiff -> m ()

PrimMonad m => Mutable m CULong

Instance details

Associated Types

type Ref m CULong = (v :: Type)

Methods

thawRef :: CULong -> m (Ref m CULong)

freezeRef :: Ref m CULong -> m CULong

copyRef :: Ref m CULong -> CULong -> m ()

PrimMonad m => Mutable m CLong

Instance details

Associated Types

type Ref m CLong = (v :: Type)

Methods

thawRef :: CLong -> m (Ref m CLong)

freezeRef :: Ref m CLong -> m CLong

copyRef :: Ref m CLong -> CLong -> m ()

PrimMonad m => Mutable m CUInt

Instance details

Associated Types

type Ref m CUInt = (v :: Type)

Methods

thawRef :: CUInt -> m (Ref m CUInt)

freezeRef :: Ref m CUInt -> m CUInt

copyRef :: Ref m CUInt -> CUInt -> m ()

PrimMonad m => Mutable m CInt

Instance details

Associated Types

type Ref m CInt = (v :: Type)

Methods

thawRef :: CInt -> m (Ref m CInt)

freezeRef :: Ref m CInt -> m CInt

copyRef :: Ref m CInt -> CInt -> m ()

PrimMonad m => Mutable m CUShort

Instance details

Associated Types

type Ref m CUShort = (v :: Type)

Methods

thawRef :: CUShort -> m (Ref m CUShort)

freezeRef :: Ref m CUShort -> m CUShort

copyRef :: Ref m CUShort -> CUShort -> m ()

PrimMonad m => Mutable m CShort

Instance details

Associated Types

type Ref m CShort = (v :: Type)

Methods

thawRef :: CShort -> m (Ref m CShort)

freezeRef :: Ref m CShort -> m CShort

copyRef :: Ref m CShort -> CShort -> m ()

PrimMonad m => Mutable m CUChar

Instance details

Associated Types

type Ref m CUChar = (v :: Type)

Methods

thawRef :: CUChar -> m (Ref m CUChar)

freezeRef :: Ref m CUChar -> m CUChar

copyRef :: Ref m CUChar -> CUChar -> m ()

PrimMonad m => Mutable m CSChar

Instance details

Associated Types

type Ref m CSChar = (v :: Type)

Methods

thawRef :: CSChar -> m (Ref m CSChar)

freezeRef :: Ref m CSChar -> m CSChar

copyRef :: Ref m CSChar -> CSChar -> m ()

PrimMonad m => Mutable m CChar

Instance details

Associated Types

type Ref m CChar = (v :: Type)

Methods

thawRef :: CChar -> m (Ref m CChar)

freezeRef :: Ref m CChar -> m CChar

copyRef :: Ref m CChar -> CChar -> m ()

PrimMonad m => Mutable m Word64

Instance details

Associated Types

type Ref m Word64 = (v :: Type)

Methods

thawRef :: Word64 -> m (Ref m Word64)

freezeRef :: Ref m Word64 -> m Word64

copyRef :: Ref m Word64 -> Word64 -> m ()

PrimMonad m => Mutable m Word16

Instance details

Associated Types

type Ref m Word16 = (v :: Type)

Methods

thawRef :: Word16 -> m (Ref m Word16)

freezeRef :: Ref m Word16 -> m Word16

copyRef :: Ref m Word16 -> Word16 -> m ()

PrimMonad m => Mutable m Word8

Instance details

Associated Types

type Ref m Word8 = (v :: Type)

Methods

thawRef :: Word8 -> m (Ref m Word8)

freezeRef :: Ref m Word8 -> m Word8

copyRef :: Ref m Word8 -> Word8 -> m ()

PrimMonad m => Mutable m Word

Instance details

Associated Types

type Ref m Word = (v :: Type)

Methods

thawRef :: Word -> m (Ref m Word)

freezeRef :: Ref m Word -> m Word

copyRef :: Ref m Word -> Word -> m ()

PrimMonad m => Mutable m Char

Instance details

Associated Types

type Ref m Char = (v :: Type)

Methods

thawRef :: Char -> m (Ref m Char)

freezeRef :: Ref m Char -> m Char

copyRef :: Ref m Char -> Char -> m ()

PrimMonad m => Mutable m Bool

Instance details

Associated Types

type Ref m Bool = (v :: Type)

Methods

thawRef :: Bool -> m (Ref m Bool)

freezeRef :: Ref m Bool -> m Bool

copyRef :: Ref m Bool -> Bool -> m ()

PrimMonad m => Mutable m Double

Instance details

Associated Types

type Ref m Double = (v :: Type)

Methods

thawRef :: Double -> m (Ref m Double)

freezeRef :: Ref m Double -> m Double

copyRef :: Ref m Double -> Double -> m ()

PrimMonad m => Mutable m Float

Instance details

Associated Types

type Ref m Float = (v :: Type)

Methods

thawRef :: Float -> m (Ref m Float)

freezeRef :: Ref m Float -> m Float

copyRef :: Ref m Float -> Float -> m ()

PrimMonad m => Mutable m Natural

Instance details

Associated Types

type Ref m Natural = (v :: Type)

Methods

thawRef :: Natural -> m (Ref m Natural)

freezeRef :: Ref m Natural -> m Natural

copyRef :: Ref m Natural -> Natural -> m ()

PrimMonad m => Mutable m Integer

Instance details

Associated Types

type Ref m Integer = (v :: Type)

Methods

thawRef :: Integer -> m (Ref m Integer)

freezeRef :: Ref m Integer -> m Integer

copyRef :: Ref m Integer -> Integer -> m ()

PrimMonad m => Mutable m Int

Instance details

Associated Types

type Ref m Int = (v :: Type)

Methods

thawRef :: Int -> m (Ref m Int)

freezeRef :: Ref m Int -> m Int

copyRef :: Ref m Int -> Int -> m ()

(PrimMonad m, Prim a) => Mutable m (Vector a)
Mutable reference is MVector.

Instance details

Associated Types

type Ref m (Vector a) = (v :: Type)

Methods

thawRef :: Vector a -> m (Ref m (Vector a))

freezeRef :: Ref m (Vector a) -> m (Vector a)

copyRef :: Ref m (Vector a) -> Vector a -> m ()

(PrimMonad m, Unbox a) => Mutable m (Vector a)
Mutable reference is MVector.

Instance details

Associated Types

type Ref m (Vector a) = (v :: Type)

Methods

thawRef :: Vector a -> m (Ref m (Vector a))

freezeRef :: Ref m (Vector a) -> m (Vector a)

copyRef :: Ref m (Vector a) -> Vector a -> m ()

(PrimMonad m, Storable a) => Mutable m (Vector a)
Mutable reference is MVector.

Instance details

Associated Types

type Ref m (Vector a) = (v :: Type)

Methods

thawRef :: Vector a -> m (Ref m (Vector a))

freezeRef :: Ref m (Vector a) -> m (Vector a)

copyRef :: Ref m (Vector a) -> Vector a -> m ()

PrimMonad m => Mutable m (Vector a)
Mutable reference is MVector.

Instance details

Associated Types

type Ref m (Vector a) = (v :: Type)

Methods

thawRef :: Vector a -> m (Ref m (Vector a))

freezeRef :: Ref m (Vector a) -> m (Vector a)

copyRef :: Ref m (Vector a) -> Vector a -> m ()

Mutable m a => Mutable m (Identity a)
Meant for usage with higher-kinded data pattern (See HKD)

Instance details

Associated Types

type Ref m (Identity a) = (v :: Type)

Methods

thawRef :: Identity a -> m (Ref m (Identity a))

freezeRef :: Ref m (Identity a) -> m (Identity a)

copyRef :: Ref m (Identity a) -> Identity a -> m ()

(PrimMonad m, Mutable m a) => Mutable m [a]
Mutable linked list with mutable references in each cell. See unconsListRef and consListRef for ways to directly work with this type as a mutable linked list.

Instance details

Associated Types

type Ref m [a] = (v :: Type)

Methods

thawRef :: [a] -> m (Ref m [a])

freezeRef :: Ref m [a] -> m [a]

copyRef :: Ref m [a] -> [a] -> m ()

(Mutable m a, PrimMonad m) => Mutable m (Maybe a)

Instance details

Associated Types

type Ref m (Maybe a) = (v :: Type)

Methods

thawRef :: Maybe a -> m (Ref m (Maybe a))

freezeRef :: Ref m (Maybe a) -> m (Maybe a)

copyRef :: Ref m (Maybe a) -> Maybe a -> m ()

Mutable m a => Mutable m (Dual a)

Instance details

Associated Types

type Ref m (Dual a) = (v :: Type)

Methods

thawRef :: Dual a -> m (Ref m (Dual a))

freezeRef :: Ref m (Dual a) -> m (Dual a)

copyRef :: Ref m (Dual a) -> Dual a -> m ()

Mutable m a => Mutable m (Down a)

Instance details

Associated Types

type Ref m (Down a) = (v :: Type)

Methods

thawRef :: Down a -> m (Ref m (Down a))

freezeRef :: Ref m (Down a) -> m (Down a)

copyRef :: Ref m (Down a) -> Down a -> m ()

Mutable m a => Mutable m (Sum a)

Instance details

Associated Types

type Ref m (Sum a) = (v :: Type)

Methods

thawRef :: Sum a -> m (Ref m (Sum a))

freezeRef :: Ref m (Sum a) -> m (Sum a)

copyRef :: Ref m (Sum a) -> Sum a -> m ()

Mutable m a => Mutable m (Product a)

Instance details

Associated Types

type Ref m (Product a) = (v :: Type)

Methods

thawRef :: Product a -> m (Ref m (Product a))

freezeRef :: Ref m (Product a) -> m (Product a)

copyRef :: Ref m (Product a) -> Product a -> m ()

Mutable m a => Mutable m (Identity a)

Instance details

Associated Types

type Ref m (Identity a) = (v :: Type)

Methods

thawRef :: Identity a -> m (Ref m (Identity a))

freezeRef :: Ref m (Identity a) -> m (Identity a)

copyRef :: Ref m (Identity a) -> Identity a -> m ()

PrimMonad m => Mutable m (Complex a)

Instance details

Associated Types

type Ref m (Complex a) = (v :: Type)

Methods

thawRef :: Complex a -> m (Ref m (Complex a))

freezeRef :: Ref m (Complex a) -> m (Complex a)

copyRef :: Ref m (Complex a) -> Complex a -> m ()

PrimMonad m => Mutable m (Ratio a)

Instance details

Associated Types

type Ref m (Ratio a) = (v :: Type)

Methods

thawRef :: Ratio a -> m (Ref m (Ratio a))

freezeRef :: Ref m (Ratio a) -> m (Ratio a)

copyRef :: Ref m (Ratio a) -> Ratio a -> m ()

(Monad m, Mutable m a, Mutable m b) => Mutable m (a, b)

Instance details

Associated Types

type Ref m (a, b) = (v :: Type)

Methods

thawRef :: (a, b) -> m (Ref m (a, b))

freezeRef :: Ref m (a, b) -> m (a, b)

copyRef :: Ref m (a, b) -> (a, b) -> m ()

(Mutable m a, Mutable m b, PrimMonad m) => Mutable m (Either a b)

Instance details

Associated Types

type Ref m (Either a b) = (v :: Type)

Methods

thawRef :: Either a b -> m (Ref m (Either a b))

freezeRef :: Ref m (Either a b) -> m (Either a b)

copyRef :: Ref m (Either a b) -> Either a b -> m ()

(Monad m, RecApplicative as, NatToInt (RLength as), RPureConstrained (IndexableField as) as, Mutable m (Rec f as), Ref m (Rec f as) ~ Rec (RecRef m f) as) => Mutable m (ARec f as)

Instance details

Associated Types

type Ref m (ARec f as) = (v :: Type)

Methods

thawRef :: ARec f as -> m (Ref m (ARec f as))

freezeRef :: Ref m (ARec f as) -> m (ARec f as)

copyRef :: Ref m (ARec f as) -> ARec f as -> m ()

(Monad m, Mutable m (f a2), Mutable m (Rec f as), Ref m (Rec f as) ~ Rec (RecRef m f) as) => Mutable m (Rec f (a2 ': as))

Instance details

Associated Types

type Ref m (Rec f (a2 ': as)) = (v :: Type)

Methods

thawRef :: Rec f (a2 ': as) -> m (Ref m (Rec f (a2 ': as)))

freezeRef :: Ref m (Rec f (a2 ': as)) -> m (Rec f (a2 ': as))

copyRef :: Ref m (Rec f (a2 ': as)) -> Rec f (a2 ': as) -> m ()

Monad m => Mutable m (Rec f ('[] :: [u]))

Instance details

Associated Types

type Ref m (Rec f '[]) = (v :: Type)

Methods

thawRef :: Rec f '[] -> m (Ref m (Rec f '[]))

freezeRef :: Ref m (Rec f '[]) -> m (Rec f '[])

copyRef :: Ref m (Rec f '[]) -> Rec f '[] -> m ()

(Monad m, Mutable m a, Mutable m b, Mutable m c) => Mutable m (a, b, c)

Instance details

Associated Types

type Ref m (a, b, c) = (v :: Type)

Methods

thawRef :: (a, b, c) -> m (Ref m (a, b, c))

freezeRef :: Ref m (a, b, c) -> m (a, b, c)

copyRef :: Ref m (a, b, c) -> (a, b, c) -> m ()

Mutable m a => Mutable m (Const a b)

Instance details

Associated Types

type Ref m (Const a b) = (v :: Type)

Methods

thawRef :: Const a b -> m (Ref m (Const a b))

freezeRef :: Ref m (Const a b) -> m (Const a b)

copyRef :: Ref m (Const a b) -> Const a b -> m ()

Mutable m a => Mutable m (Const a b)

Instance details

Associated Types

type Ref m (Const a b) = (v :: Type)

Methods

thawRef :: Const a b -> m (Ref m (Const a b))

freezeRef :: Ref m (Const a b) -> m (Const a b)

copyRef :: Ref m (Const a b) -> Const a b -> m ()

(Monad m, Mutable m a, Mutable m b, Mutable m c, Mutable m d) => Mutable m (a, b, c, d)

Instance details

Associated Types

type Ref m (a, b, c, d) = (v :: Type)

Methods

thawRef :: (a, b, c, d) -> m (Ref m (a, b, c, d))

freezeRef :: Ref m (a, b, c, d) -> m (a, b, c, d)

copyRef :: Ref m (a, b, c, d) -> (a, b, c, d) -> m ()

(Mutable m (f a), Mutable m (g a), PrimMonad m) => Mutable m (Sum f g a)

Instance details

Associated Types

type Ref m (Sum f g a) = (v :: Type)

Methods

thawRef :: Sum f g a -> m (Ref m (Sum f g a))

freezeRef :: Ref m (Sum f g a) -> m (Sum f g a)

copyRef :: Ref m (Sum f g a) -> Sum f g a -> m ()

(Mutable m (f a), Mutable m (g a)) => Mutable m (Product f g a)

Instance details

Associated Types

type Ref m (Product f g a) = (v :: Type)

Methods

thawRef :: Product f g a -> m (Ref m (Product f g a))

freezeRef :: Ref m (Product f g a) -> m (Product f g a)

copyRef :: Ref m (Product f g a) -> Product f g a -> m ()

Mutable m (f (g a)) => Mutable m (Compose f g a)

Instance details

Associated Types

type Ref m (Compose f g a) = (v :: Type)

Methods

thawRef :: Compose f g a -> m (Ref m (Compose f g a))

freezeRef :: Ref m (Compose f g a) -> m (Compose f g a)

copyRef :: Ref m (Compose f g a) -> Compose f g a -> m ()