Data.Mutable.MutPart

Description

Tools for working with individual components of piecewise-mutable values.

Documentation

newtype MutPart m s a

A MutPart m s a is a way to "zoom into" an a, as a part of a mutable reference on s. This allows you to only modify a single a part of the s, without touching the rest.

An example that is commonly found in the ecosystem is something like (flipped) write :: Int -> MVector s a -> a -> m () from Data.Vector.Mutable --- write 3 :: MVector s a -> a -> m (), for instance, lets you modify a specific part of the vector without touching the rest.

You would use a MutPart using freezePart, copyPart, modifyPart, etc.

For non-composite types, there won't really be any meaningful values. However, we have them for many composite types. For example, for tuples:

mutFst :: MutPart m (a, b) a
mutSnd :: MutPart m (a, b) b

ghci> r <- thawRef (2, 4)
ghci> copyPart mutFst r 100
ghci> freezeRef r
(100, 4)

If you are using GRef as an automatically-defined mutable reference, then the easiest way to create these for your mutable types are with fieldMut and posMut.

If you are using the "Higher-kinded data" pattern, then there's an easy way to generate a MutPart for every single field, if you have a product type --- see hkdMutParts for more information.

Constructors

MutPart

Fields

getMutPart :: Ref m s -> Ref m a

withMutPart :: MutPart m s a -> Ref m s -> (Ref m a -> m b) -> m b

Using a MutPart, perform a function on a Ref m s as if you had a Ref m a.

freezePart :: Mutable m a => MutPart m s a -> Ref m s -> m a

With a MutPart, read out a specific part of a Ref.

copyPart :: Mutable m a => MutPart m s a -> Ref m s -> a -> m ()

With a MutPart, overwrite into a specific part of a Ref.

modifyPart :: Mutable m a => MutPart m s a -> Ref m s -> (a -> a) -> m ()

With a MutPart, modify a specific part of a Ref with a pure function.

modifyPart' :: Mutable m a => MutPart m s a -> Ref m s -> (a -> a) -> m ()

modifyPart, but forces the result before storing it back in the reference.

updatePart :: Mutable m a => MutPart m s a -> Ref m s -> (a -> (a, b)) -> m b

updateRef, under a MutPart to only modify a specific part of a Ref.

updatePart' :: Mutable m a => MutPart m s a -> Ref m s -> (a -> (a, b)) -> m b

updatePart, but forces the result before storing it back in the reference.

compMP :: MutPart m a b -> MutPart m b c -> MutPart m a c infixr 9

Compose two MutParts one after the other.

Note this is also available (albeit flipped in arguments) through the Category instance.

idMP :: MutPart m a a

The identity MutPart: simply focus into the same type itself.

Note this is also available through the Category instance.

Built-in MutPart

mutFst :: MutPart m (a, b) a

MutPart into the first field of a tuple reference.

mutSnd :: MutPart m (a, b) b

MutPart into the second field of a tuple reference.

class (Mutable m s, Mutable m a) => FieldMut (fld :: Symbol) m s a | fld s -> a

Create a MutPart for a field name. Should work for any type with one constructor whose mutable reference is GRef. See fieldMut for usage directions.

Minimal complete definition

fieldMut

fieldMut :: FieldMut fld m s a => Label fld -> MutPart m s a

Create a MutPart for a field name. Should work for any type with one constructor whose mutable reference is GRef.

Is meant to be used with OverloadedLabels:

data Foo = Foo { fInt :: Int, fDouble :: Double }
  deriving (Generic, Show)

instance Mutable m Foo where
    type Ref m Foo = GRef m Foo

ghci> r <- thawRef (Foo 3 4.5)
ghci> freezePart (fieldMut #fInt) r
3
ghci> copyPart (fieldMut #fDouble) 1.23
ghci> freezeRef r
Foo 3 1.23

However, you can use it without OverloadedLabels by using Label with TypeApplications:

ghci> freezePart (fieldMut (Label @"fInt")) r
3

This and posMut are the main ways to generate a MutPart for a type whose mutable reference is GRef. Note that because all of the lookups are done at compile-time, fieldMut and posMut have more or less identical performance characteristics.

data Label (a :: Symbol)

Proxy for label type

Constructors

Label

class (Mutable m s, Mutable m a) => PosMut (i :: Nat) m s a | i s -> a

Create a MutPart for a position in a sum type. Should work for any type with one constructor whose mutable reference is GRef. See posMut for usage directions.

Minimal complete definition

posMut

posMut :: PosMut i m s a => MutPart m s a

Create a MutPart for a position in a sum type. Should work for any type with one constructor whose mutable reference is GRef.

Meant to be used with TypeApplications:

data Foo = Foo Int Double
  deriving (Generic, Show)

instance Mutable m Foo where
    type Ref m Foo = GRef m Foo

ghci> r <- thawRef (Foo 3 4.5)
ghci> freezePart (posMut 1) r
3
ghci> copyPart (posMut 2) 1.23
ghci> freezeRef r
Foo 3 1.23

This and fieldMut are the main ways to generate a MutPart for a type whose mutable reference is GRef. Note that because all of the lookups are done at compile-time, posMut and fieldMut have more or less identical performance characteristics.

hkdMutParts :: forall z m. (Generic (z (RefFor m)), Generic (z (MutPart m (z Identity))), HKDMutParts m z (Rep (z (RefFor m))) (Rep (z (MutPart m (z Identity))))) => z (MutPart m (z Identity))

If you are using the "higher-kinded data" pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/, and you have the appropriate instance for Ref, then you can use this to generate a MutPart for every field, if you have a type with only one constructor.

data MyTypeF f = MT
     { fInt    :: f Int
     , fDouble :: f Double
     }
  deriving Generic

instance Mutable (MyTypeF Identity) where
    type Ref (MyTypeF Identity) = MyTypeF (RefFor m)

mx :: MutPart (MyTypeF Identity) (Vector Int)
my :: MutPart (MyTypeF Identity) (V.Vector Double)
MT mx my = hkdMutParts @MyTypeF

ghci> r <- thawRef (MT 3 4.5)
ghci> freezePart mx r
3
ghci> copyPart (fDouble (hkdMutParts @MyTypeF)) r 12.3
ghci> freezeRef r
MT 3 12.3

class (Mutable m (z Identity), Ref m (z Identity) ~ z (RefFor m)) => HKDMutParts m z i o

Typeclass used to implement hkdMutParts. See documentation of hkdMutParts for more information.

Minimal complete definition

hkdMutParts_

mutRec :: forall a as f rec m. (Ref m (rec f as) ~ rec (RecRef m f) as, RecElem rec a a as as (RIndex a as), RecElemFCtx rec (RecRef m f)) => MutPart m (rec f as) (f a)

A MutPart for a field in a vinyl Rec, automatically generated as the first field with a matching type. This is polymorphic to work over both Rec and ARec.

ghci> r <- thawRef $ [1,2,3] :& [True, False] :& RNil
ghci> modifyPart (mutRec @Bool) r reverse
ghci> freezeRef r
[1,2,3] :& [False, True] :& RNil

coerceRef :: Ref m s ~ CoerceRef m s a => MutPart m s a

A MutPart to get into a CoerceRef.