Data.Mutable.Class

Description

Provides the Mutable typeclass and various helpers. See Mutable for the main "entrypoint".

Documentation

class Monad m => Mutable m a where

An instance of Mutable m a means that a can be stored a mutable reference in monad m.

The associated type Ref m a links any a to the type of its canonical mutable version.

The benefit of this typeclass, instead of just using IORef or MutVar or specific mutable versions like Vector and MVector, is two-fold:

Piecewise-mutable values, so you can write to only one part and not others. This also allows for cheaper "writes", even if you replace the whole value: you don't need to ever synthesize an entire new value, you can keep each component in a separate variable until you freezeRef it out. This can be especially useful for composite data types containing large structures like Vector.

Generic abstractions (similar to Show), so you can automatically derive instances while preserving piecewise-ness. For example, the instance

instance (Mutable m a, Mutable m b) => Mutable m (a, b)

If a and b are piecwise-mutable, then the instance here will appropriately utilize that fact.

To modify the specific parts of mutable values, it can be useful to use the functions in Data.Mutable.MutPart.

There are facilities to automatically piecewise mutable versions for user-defined instances of Generic.

For example, if we have a type like:

data TwoVectors = TV
    { tvInt    :: Vector Int
    , tvDouble :: Vector Double
    }
  deriving Generic

instance Mutable m TwoVectors where
    type Ref m Foo = GRef m TwoVectors

Then now we get:

thawRef   :: TwoVectors -> m (GRef m TwoVectors)
freezeRef :: GRef m TwoVectors -> m TwoVectors

And GRef m TwoVectors is now a piecewise-mutable reference storing each part in a way that can be modified separately (for example, with tools from Data.Mutable.MutPart). It does this by internally allocating two MVectors. If the two vectors are large, this can be much more efficient to modify (if you are modifying several times) than by just doing alterations on TwoVectors. It is also much better for large vectors if you plan on modifying only a single item in the vector.

If you are using the "higher-kinded" data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/, then we can also do:

data TwoVectors f = TV
     { tvInt    :: HKD f (Vector Int)
     , tvDouble :: HKD f (Vector Double)
     }
  deriving Generic

instance Mutable (TwoVectors Identity) where
    type Ref (TwoVectors Identity) = TwoVectors (RefFor m)

And now your mutable ref is literally going to be a product of the components

ghci> tvr@(TV is ds) <- thawRef (TV xs ys)
ghci> :t tvr
TV (RefFor IO)
ghci> :t is
MVector RealWorld Int
ghci> :t ds
MVector RealWorld Double

So thawRef will actually just get you the same record type but with the mutable versions of each field. If you modify the mutable fields, and then later freezeRef the whole thing, the resulting frozen value will incorporate all of the changes to the individual fields.

In addition, there are two more "automatically derived" instances you can get by picking Ref:

-- Make a mutable version for any newtype wrapper, using the Mutable
-- of the underlying type
newtype MyType = MT (Vector Double)

type Ref m MyType = CoerceRef m MyType (Vector Double)
 
-- Make a mutable version of any container, where the items are all
-- mutable references.
data MyContainer a = MC a a a a
  deriving (Functor, Foldable, Traversable)

type Ref m (MyContainer a) = TraverseRef m MyContainer a

Minimal complete definition

Nothing

Associated Types

type Ref m a = (v :: Type) | v -> a

Links the type a to the type of its canonical mutable version.

For example, for Vector, the mutable version is MVector, so we have

type Ref m (Vector a) = MVector (PrimState m) a

This means that using thawRef on a Vector will give you an MVector, using freezeRef on a Vector will give you a Vector, etc.

thawRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector a
    -> m (Vector s a)

freezeRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> m (Vector a)

copyRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> Vector a
    -> m ()

This associated type must be unique for a, so no two types a can have the same Ref m a. This makes type inference a lot more useful: if you use freezeRef on an MVector, for instance, the return type will be inferred to be Vector.

The default instance is just a plain old MutVar containing the type. This is a valid instance, but it treats the entire type "wholesale" --- it is basically using it as a non-mutable type. You won't get any of the performance benefits of piecewise mutation from it, but it is useful as a base case for non-composite types like Int.

There are some built-in alternative options for user-defined ADTs with Generic instances:

-- Works for all Generic instances, preserves piecewise mutation
-- for products
type Ref m a = GRef m a

-- Works for "higher-kinded" data types, a la
-- https://reasonablypolymorphic.com/blog/higher-kinded-data/
type Ref m (z Identity) = z (RefFor m)

If you just set up a blank instance, the implementations of thawRef, freezeRef, and copyRef will be inferred using DefaultMutable.

data MyType

-- The default setup is OK
instance Mutable m MyType

-- This is equivalent to the above
instance Mutable m MyType
    type Ref m MyType = MutVar (PrimState m) MyType

-- any Generic instance
data Foo = Foo { fInt :: Int, fDouble :: Double }
  deriving Generic

instance Mutable m Foo where
    type Ref m Foo = GRef m Foo

-- HKD pattern types
data Bar f = Bar { bInt :: f Int, bDouble :: f Double }
  deriving Generic

instance Mutable (Bar Identity) where
    type Ref (Bar Identity) = Bar (RefFor m)

type Ref m a = MutVar (PrimState m) a

Methods

thawRef :: a -> m (Ref m a)

Thaw a pure/persistent value into its mutable version, which can be manipulated using modifyRef or other methods specific for that type (like read).

Returns the Ref instance, so, for example, for Vector:

thawRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector a
    -> m (Vector s a)

For non-composite (like Int), this is often called the "new var" function, like newIORef / newSTRef / newMutVar etc.

default thawRef :: DefaultMutable m a (Ref m a) => a -> m (Ref m a)

freezeRef :: Ref m a -> m a

Freeze a mutable value into its pure/persistent version.

Takes a Ref instance, but type inference will be able to infer the pure value's type because Ref is injective.

For example, for Vector:

freezeRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> m (Vector a)

For non-composite (like Int), this is often called the "read var" function, like readIORef / readSTRef / readMutVar etc.

default freezeRef :: DefaultMutable m a (Ref m a) => Ref m a -> m a

copyRef :: Ref m a -> a -> m ()

Overwrite a mutable value by provivding a pure/persistent value. copyRef

Returns the Ref and the value, so, for example, for Vector:

copyRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> Vector a
    -> m ()

Note that if a is a composite type (with an appropriate composite reference), this will be done "piecewise": it'll write to each mutable component separately.

For non-composite (like Int), this is often called the "write var" function, like writeIORef / writeSTRef / writeMutVar etc.

default copyRef :: DefaultMutable m a (Ref m a) => Ref m a -> a -> m ()

modifyRef :: Mutable m a => Ref m a -> (a -> a) -> m ()

Apply a pure function on an immutable value onto a value stored in a mutable reference.

modifyRef' :: Mutable m a => Ref m a -> (a -> a) -> m ()

modifyRef, but forces the result before storing it back in the reference.

updateRef :: Mutable m a => Ref m a -> (a -> (a, b)) -> m b

Apply a pure function on an immutable value onto a value stored in a mutable reference, returning a result value from that function.

updateRef' :: Mutable m a => Ref m a -> (a -> (a, b)) -> m b

updateRef, but forces the updated value before storing it back in the reference.

newtype RefFor m a

A handy newtype wrapper that allows you to partially apply Ref. RefFor m a is the same as Ref m a, but can be partially applied.

If used with HKD, you can treat this syntactically identically as a Ref m a.

Constructors

RefFor

Fields

getRefFor :: Ref m a

class DefaultMutable m a r where

The default implementations of thawRef, freezeRef, and copyRef dispatched for different choices of Ref.

Basically, by specifying Ref, you get the rest of the instance for free.

We have the default case:

-- default, if you don't specify Ref
instance Mutable m MyType

-- the above is the same as:
instance Mutable m MyType
    type Ref m MyType = MutVar (PrimState m) MyType

The case for any instance of Generic:

instance Mutable m MyType
    type Ref m MyType = GRef m MyType

The case for the "higher-kinded data" pattern a la https://reasonablypolymorphic.com/blog/higher-kinded-data/:

instance Mutable m (MyTypeF Identity)
    type Ref m (MyTypeF Identity) = MyTypeF (RefFor m)

The case for any newtype wrapper:

newtype MyType = MT (Vector Double)

instance Mutable m MyType where
    type Ref m MyType = CoerceRef m MyType (Vector Double)

And the case for any 'Traversable instance, where the items will all be mutable references:

data MyContainer a = MC a a a a
  deriving (Functor, Foldable, Traversable)

instance Mutable m a => Mutable m (MyContainer a) where
    type Ref m (MyContainer a) = TraverseRef m MyContainer a

Methods

defaultThawRef :: a -> m r

defaultFreezeRef :: r -> m a

defaultCopyRef :: r -> a -> m ()

Providing and overwriting instances

newtype VarMut a

Newtype wrapper that can provide any type with a Mutable instance, giving it a "non-piecewise" instance. Can be useful for avoiding orphan instances yet still utilizing auto-deriving features, or for overwriting the Mutable instance of other instances.

For example, let's say you want to auto-derive an instance for your data type:

data MyType = MT Int Double OtherType
  deriving Generic

This is possible if all of MyTypes fields have Mutable instances. However, let's say OtherType comes from an external library that you don't have control over, and so you cannot give it a Mutable instance without incurring an orphan instance.

One solution is to wrap it in VarMut:

data MyType = MT Int Double (VarMut OtherType)
  deriving Generic

This can then be auto-derived:

instance Mutable m MyType where
    type Ref m MyType = GRef m MyType

It can also be used to override a Mutable instance. For example, even if the Mutable instance of SomeType is piecewise-mutable, the Mutable instance of VarMut SomeType will be not be piecewise.

For example, the Mutable instance for String is a mutable linked list, but it might be more efficient to treat it as an atomic value to update all at once. You can use VarMut String to get that Mutable instance.

Constructors

VarMut

Fields

getVarMut :: a

newtype CoerceMut s a

Similar to VarMut, this allows you to overwrite the normal Mutable instance of a type to utilize a coercible type's Mutable instance instead of its normal instance. It's also useful to provide an instance for an externally defined type without incurring orphan instances.

For example, if an external library provides

newtype DoubleVec = DV (Vector Double)

and you want to use it following Vectors Mutable instance (via MVector), but you don't want to write an orphan instance like

instance Mutable m DoubleVec where
    type Ref m DoubleVec = CoerceRef m DoubleVec (Vector Double)

then you can instead use CoerceMut DoubleVec (Vector Double) as the data type. This wrapped type does use the inderlying Mutable insatnce for Vector.

Constructors

CoerceMut

Fields

getCoerceMut :: s

newtype TraverseMut f a

Similar to VarMut, this allows you to overwrite the normal Mutable instance for a type to utilize its Traversable instance instead of its normal instance. It's also useful to provide an instance for an externally defined type without incurring orphan instances.

For example, the instance of Mutable (TraverseMut [] a) is a normal list of mutable references, instead of a full-on mutable linked list.

Constructors

TraverseMut

Fields

getTraverseMut :: f a

newtype Immutable a

Similar to VarMut, this allows you to overwrite the normal Mutable instance of a type to make it immutable.

For example, let's say you have a type, with the automatically derived generic instance of Mutable:

data MyType = MT
    { mtX :: Int
    , mtY :: Vector Double
    , mtZ :: String
    }
  deriving Generic

instance Mutable m MyType where
    type Ref m MyType = GRef m MyType

This basically uses three mutable references: the Int, the Vector Double, and the String. However, you might want the Mutable instance of MyType to be immutable String field, and so it cannot be updated at all even when thawed. To do that, you can instead have:

data MyType = MT
    { mtX :: Int
    , mtY :: Vector Double
    , mtZ :: Immutable String
    }
  deriving Generic

instance Mutable m MyType where
    type Ref m MyType = GRef m MyType

which has that behavior. The Int and the Vector will be mutable within Ref m MyType, but not the String.

Constructors

Immutable

Fields

getImmutable :: a

Changing underlying monad

reMutable :: forall m n a r. (Mutable m a, Monad n) => (forall x. m x -> n x) -> (Mutable n a => r) -> r

If you can provice a natural transformation from m to n, you should be able to use a value as if it had Mutable n a if you have Mutable m a.

reMutableConstraint :: forall m n a. (Mutable m a, Monad n) => (forall x. m x -> n x) -> Mutable m a :- Mutable n a

If you can provice a natural transformation from m to n, then Mutable m a should also imply Mutable n a.