Data.Mutable.Class

Description

Provides the Mutable typeclass and various helpers. See Mutable for the main "entrypoint".

Documentation

class Monad m => Mutable m a where

An instance of Mutable m a means that a can be stored a mutable reference in monad m.

The associated type Ref m a links any a to the type of its canonical mutable version.

The benefit of this typeclass, instead of just using IORef or MutVar or specific mutable versions like Vector and MVector, is two-fold:

Piecewise-mutable values, so you can write to only one part and not others. This also allows for cheaper "writes", even if you replace the whole value: you don't need to ever synthesize an entire new value, you can keep each component in a separate variable until you freezeRef it out. This can be especially useful for composite data types containing large structures like Vector.

Generic abstractions (similar to Show), so you can automatically derive instances while preserving piecewise-ness. For example, the instance

instance (Mutable m a, Mutable m b) => Mutable m (a, b)

If a and b are piecwise-mutable, then the instance here will appropriately utilize that fact.

To modify the specific parts of mutable values, it can be useful to use the functions in Data.Mutable.MutPart.

There are facilities to automatically piecewise mutable versions for user-defined instances of Generic.

For example, if we have a type like:

data TwoVectors = TV
    { tvInt    :: Vector Int
    , tvDouble :: Vector Double
    }
  deriving Generic

instance Mutable m TwoVectors where
    type Ref m Foo = GRef m TwoVectors

Then now we get:

thawRef   :: TwoVectors -> m (GRef m TwoVectors)
freezeRef :: GRef m TwoVectors -> m TwoVectors

And GRef m TwoVectors is now a piecewise-mutable reference storing each part in a way that can be modified separately (for example, with tools from Data.Mutable.MutPart). It does this by internally allocating two MVectors. If the two vectors are large, this can be much more efficient to modify (if you are modifying several times) than by just doing alterations on TwoVectors. It is also much better for large vectors if you plan on modifying only a single item in the vector.

If you are using the "higher-kinded" data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/, then we can also do:

data TwoVectors f = TV
     { tvInt    :: HKD f (Vector Int)
     , tvDouble :: HKD f (Vector Double)
     }
  deriving Generic

instance Mutable (TwoVectors Identity) where
    type Ref (TwoVectors Identity) = TwoVectors (RefFor m)

And now your mutable ref is literally going to be a product of the components

ghci> tvr@(TV is ds) <- thawRef (TV xs ys)
ghci> :t tvr
TV (RefFor IO)
ghci> :t is
MVector RealWorld Int
ghci> :t ds
MVector RealWorld Double

So thawRef will actually just get you the same record type but with the mutable versions of each field. If you modify the mutable fields, and then later freezeRef the whole thing, the resulting frozen value will incorporate all of the changes to the individual fields.

Minimal complete definition

Nothing

Associated Types

type Ref m a = (v :: Type) | v -> a

Links the type a to the type of its canonical mutable version.

For example, for Vector, the mutable version is MVector, so we have

type Ref m (Vector a) = MVector (PrimState m) a

This means that using thawRef on a Vector will give you an MVector, using freezeRef on a Vector will give you a Vector, etc.

thawRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector a
    -> m (Vector s a)

freezeRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> m (Vector a)

copyRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> Vector a
    -> m ()

This associated type must be unique for a, so no two types a can have the same Ref m a. This makes type inference a lot more useful: if you use freezeRef on an MVector, for instance, the return type will be inferred to be Vector.

The default instance is just a plain old MutVar containing the type. This is a valid instance, but it treats the entire type "wholesale" --- it is basically using it as a non-mutable type. You won't get any of the performance benefits of piecewise mutation from it, but it is useful as a base case for non-composite types like Int.

There are some built-in alternative options for user-defined ADTs with Generic instances:

-- Works for all Generic instances, preserves piecewise mutation
-- for products
type Ref m a = GRef m a

-- Works for "higher-kinded" data types, a la
-- https://reasonablypolymorphic.com/blog/higher-kinded-data/
type Ref m (z Identity) = z (RefFor m)

If you just set up a blank instance, the implementations of thawRef, freezeRef, and copyRef will be inferred using DefaultMutable.

data MyType

-- The default setup is OK
instance Mutable m MyType

-- This is equivalent to the above
instance Mutable m MyType
    type Ref m MyType = MutVar (PrimState m) MyType

-- any Generic instance
data Foo = Foo { fInt :: Int, fDouble :: Double }
  deriving Generic

instance Mutable m Foo where
    type Ref m Foo = GRef m Foo

-- HKD pattern types
data Bar f = Bar { bInt :: f Int, bDouble :: f Double }
  deriving Generic

instance Mutable (Bar Identity) where
    type Ref (Bar Identity) = Bar (RefFor m)

type Ref m a = MutVar (PrimState m) a

Methods

thawRef :: a -> m (Ref m a)

Thaw a pure/persistent value into its mutable version, which can be manipulated using modifyRef or other methods specific for that type (like read).

Returns the Ref instance, so, for example, for Vector:

thawRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector a
    -> m (Vector s a)

For non-composite (like Int), this is often called the "new var" function, like newIORef / newSTRef / newMutVar etc.

default thawRef :: DefaultMutable m a (Ref m a) => a -> m (Ref m a)

freezeRef :: Ref m a -> m a

Freeze a mutable value into its pure/persistent version.

Takes a Ref instance, but type inference will be able to infer the pure value's type because Ref is injective.

For example, for Vector:

freezeRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> m (Vector a)

For non-composite (like Int), this is often called the "read var" function, like readIORef / readSTRef / readMutVar etc.

default freezeRef :: DefaultMutable m a (Ref m a) => Ref m a -> m a

copyRef :: Ref m a -> a -> m ()

Overwrite a mutable value by provivding a pure/persistent value. copyRef

Returns the Ref and the value, so, for example, for Vector:

copyRef
    :: (PrimMonad m, s ~ PrimState m)
    => Vector s a
    -> Vector a
    -> m ()

Note that if a is a composite type (with an appropriate composite reference), this will be done "piecewise": it'll write to each mutable component separately.

For non-composite (like Int), this is often called the "write var" function, like writeIORef / writeSTRef / writeMutVar etc.

default copyRef :: DefaultMutable m a (Ref m a) => Ref m a -> a -> m ()

modifyRef :: Mutable m a => Ref m a -> (a -> a) -> m ()

Apply a pure function on an immutable value onto a value stored in a mutable reference.

modifyRef' :: Mutable m a => Ref m a -> (a -> a) -> m ()

modifyRef, but forces the result before storing it back in the reference.

updateRef :: Mutable m a => Ref m a -> (a -> (a, b)) -> m b

Apply a pure function on an immutable value onto a value stored in a mutable reference, returning a result value from that function.

updateRef' :: Mutable m a => Ref m a -> (a -> (a, b)) -> m b

updateRef, but forces the updated value before storing it back in the reference.

newtype MutRef a

Newtype wrapper that can provide any type with a Mutable instance, giving it a "non-piecewise" instance. Can be useful for avoiding orphan instances yet still utilizing auto-deriving features, or for overwriting the Mutable instance of other instances.

For example, let's say you want to auto-derive an instance for your data type:

data MyType = MT Int Double OtherType
  deriving Generic

This is possible if all of MyTypes fields have Mutable instances. However, let's say OtherType comes from an external library that you don't have control over, and so you cannot give it a Mutable instance without incurring an orphan instance.

One solution is to wrap it in MutRef:

data MyType = MT Int Double (MutRef OtherType)
  deriving Generic

This can then be auto-derived:

instance Mutable m MyType where
    type Ref m MyType = GRef m MyType

It can also be used to override a Mutable instance. For example, even if the Mutable instance of SomeType is piecewise-mutable, the Mutable instance of MutRef SomeType will be not be piecewise.

Constructors

MutRef

Fields

getMutRef :: a

newtype RefFor m a

A handy newtype wrapper that allows you to partially apply Ref. RefFor m a is the same as Ref m a, but can be partially applied.

If used with HKD, you can treat this syntactically identically as a Ref m a.

Constructors

RefFor

Fields

getRefFor :: Ref m a

class DefaultMutable m a r where

The default implementations of thawRef, freezeRef, and copyRef dispatched for different choices of Ref.

Basically, by specifying Ref, you get the rest of the instance for free.

-- default, if you don't specify Ref
instance Mutable m MyType

-- the above is the same as:
instance Mutable m MyType
    type Ref m MyType = MutVar (PrimState m) MyType

-- or if we have an instance of Generic:
instance Mutable m MyType
    type Ref m MyType = GRef m MyType

-- or, using the higher-kinded data pattern, like
-- https://reasonablypolymorphic.com/blog/higher-kinded-data/
instance Mutable m (MyTypeF Identity)
    type Ref m (MyTypeF Identity) = MyTypeF (RefFor m)

Methods

defaultThawRef :: a -> m r

defaultFreezeRef :: r -> m a

defaultCopyRef :: r -> a -> m ()

Instances

Generic

newtype GRef m a

Automatically generate a piecewise mutable reference for any Generic instance.

-- | any Generic instance
data Foo = Foo { fInt :: Int, fDouble :: Double }
  deriving (Generic, Show)

instance Mutable m Foo where
    type Ref m Foo = GRef m Foo

ghci> r <- thawRef (Foo 3 4.5)
ghci> freezeRef r
Foo 3 4.5
ghci> freezePart (fieldMut #fInt) r
3
ghci> copyPart (fieldMut #fDouble) 1.23
ghci> freezeRef r
Foo 3 1.23

Note that this is basically just a bunch of tupled refs for a product type. For a sum type (with multiple constructors), an extra layer of indirection is added to account for the dynamically changable shape.

See fieldMut/posMut for ways to inspect and mutate the internals of this type (as demonstrated above).

Constructors

GRef

Fields

unGRef :: GRef_ m (Rep a) ()

gThawRef :: (Generic a, GMutable m (Rep a)) => a -> m (GRef m a)

Default thawRef for GRef.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with GRef as the Ref. However, it can be useful if you are using a GRef m a just as a normal data type, independent of the Ref class. See documentation for GRef for more information.

gFreezeRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> m a

Default freezeRef for GRef.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with GRef as the Ref. However, it can be useful if you are using a GRef m a just as a normal data type, independent of the Ref class. See documentation for GRef for more information.

gCopyRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> a -> m ()

Default copyRef for GRef.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with GRef as the Ref. However, it can be useful if you are using a GRef m a just as a normal data type, independent of the Ref class. See documentation for GRef for more information.

class Monad m => GMutable m f

Class for automatic generation of Ref for Generic instances. See GRef for more information.

Minimal complete definition

gThawRef_, gFreezeRef_, gCopyRef_

Associated Types

type GRef_ m f = (u :: Type -> Type) | u -> f

Higher-Kinded Data Pattern

thawHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z Identity -> m (z (RefFor m))

Default copyRef for the higher-kinded data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with z (RefFor m) as the Ref. However, it can be useful if you are using a z (RefFor m) just as a normal data type, independent of the Ref class. See documentation for Mutable for more information.

freezeHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z (RefFor m) -> m (z Identity)

Default freezeRef for the higher-kinded data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with z (RefFor m) as the Ref. However, it can be useful if you are using a z (RefFor m) just as a normal data type, independent of the Ref class. See documentation for Mutable for more information.

copyHKD :: forall z m. (Generic (z Identity), Generic (z (RefFor m)), GMutable m (Rep (z Identity)), GRef_ m (Rep (z Identity)) ~ Rep (z (RefFor m))) => z (RefFor m) -> z Identity -> m ()

Default copyRef for the higher-kinded data pattern, a la https://reasonablypolymorphic.com/blog/higher-kinded-data/.

You likely won't ever use this directly, since it is automatically provided if you have a Mutable instance with z (RefFor m) as the Ref. However, it can be useful if you are using a z (RefFor m) just as a normal data type, independent of the Ref class. See documentation for Mutable for more information.

Miscellaneous

newtype RecRef m f a

Ref for components in a vinyl Rec.

Constructors

RecRef

Fields

getRecRef :: Ref m (f a)

Changing underlying monad

reMutable :: forall m n a r. (Mutable m a, Monad n) => (forall x. m x -> n x) -> (Mutable n a => r) -> r

If you can provice a natural transformation from m to n, you should be able to use a value as if it had Mutable n a if you have Mutable m a.

reMutableConstraint :: forall m n a. (Mutable m a, Monad n) => (forall x. m x -> n x) -> Mutable m a :- Mutable n a

If you can provice a natural transformation from m to n, then Mutable m a should also imply Mutable n a.